add deck selector in 'add problem' window
The logic for showing a deck selector with the current list of decks in the db
in 'add problem' window has been added.
To avoid checking every time the DB for that info, a simple class 'DeckMem'
(in ui_uitls) has been created. An instance of that class stores the current
list of decks in db, and has a method update() that has to be called everytime
the list in the DB is updated through the gui.
TODO:
- The integration between 'DeckListWidget' and the instance of 'DeckMen'
  must be added.