add implementation of the 'delete deck' feature, update 'deck selector'
The logic for displaying decks in the 'add problem' window has been changed.
Every time a deck is deleted or added a flag indicating the need to
update is modified. The flag in question is a Dict object that has a key
for every piece of code that needs to know if an update of decks' data is
necessary. An instantiation of a class (the solution in the previous
commit) is not necessary anymore.

A partial clean up has been performed for the logic behind the
displaying of decks in the main window.